+++ include/mbgl/shaders/vulkan/raster.hpp
@@ -17,8 +17,8 @@ struct ShaderSource<BuiltIn::RasterShader, gfx::Backend::Type::Vulkan> {
 
     static constexpr auto vertex = R"(
 
-layout(location = 0) in vec2 in_position;
-layout(location = 1) in vec2 in_texture_position;
+layout(location = 0) in ivec2 in_position;
+layout(location = 1) in ivec2 in_texture_position;
 
 layout(set = 0, binding = 1) uniform RasterDrawableUBO {
     mat4 matrix;
